(roles-permisos): restringir acceso a rutas según permisos asignados

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
 
 
 class UserController extends Controller
@@ -15,6 +16,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        Gate::authorize('read_user');
     return view('admin.users.index');
     }
 
@@ -23,6 +25,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create_user');
         $roles= Role::all();
     return view('admin.users.create', compact('roles'));
     }
@@ -32,6 +35,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create_user');
        $data= $request->validate([
        'name'     => 'required|string|max:255',
        'email'    => 'required|string|email|max:255|unique:users',
@@ -71,6 +75,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        Gate::authorize('read_user');
     return view('admin.users.show', compact('user'));
     }
 
@@ -79,6 +84,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        Gate::authorize('update_user');
         $roles= Role::all();
     return view('admin.users.edit', compact('user', 'roles'));
     }
@@ -88,6 +94,7 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        Gate::authorize('update_user');
         $data= $request->validate([
        'name'     => 'required|string|max:255',
        'email'    => 'required|string|email|max:255|unique:users,email,'.$user->id,
@@ -122,6 +129,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+         Gate::authorize('delete_user');
         $user->roles()->detach();
         $user->delete();
